estudo

+++ src/main/java/App/Controller/CarroController.java
@@ -25,7 +25,7 @@ public class CarroController {
     }
 
     @GetMapping("/findById/{id}")
-    public ResponseEntity<Carro> findById(@PathVariable int id){
+    public ResponseEntity<Carro> findById(@PathVariable Long id){
         try {
             Carro carro = this.carroService.findById(id);
             return new ResponseEntity<>(carro,HttpStatus.OK);
